Change folder structure for Library Manager compatibility
Arduino IDE Library Manager requires that the library is located in the
root of the repository.
From https://github.com/arduino/Arduino/wiki/Library-Manager-FAQ:
>Your repo is OK if its root folder contains file library.properties and
the rest of the library source code. If your library is stored in a
subfolder, your repo is not OK and we can't include your library in the
list.

This will also allow the library to be installed via Sketch > Include
Library > Add .ZIP Library...